AI Aircraft: Handle gate positions better & fix pushback
Whenever we have a gate with a route from it, we would disregard
the heading information of the gate and set the initial heading and
speed based on the first leg of the route away from it. This made all
AI aircraft start a little bit on their away-from-parking taxi when
they spawned. Pushback also had a tendency to just stop because the
aircraft never reached their pushback point.

This patch makes the aircraft check if they are at the gate, and
use the gate's heading + zero speed as initial starting points. It
also fixes the stuck pushback, although the reason for having to
disable the taxiclearance part is unknown. (Durk did the same change
for the other parts of the code in 2011.)