AI Aircraft: Don't spawn aircraft departing in >24h
We would spawn every aircraft matching the airport with a departure
time in the future. With weekly scheduling, this fills up the
airport & parking positions with lots of aircraft that will just
sit there, and does not leave room aircraft that actually will
depart soon.

This patch simply checks if the departuretime is more than 24h
away, and avoids spawning the aircraft if it is.